backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
numpy==1.24.3
faiss-cpu
requests==2.31.0
python-dotenv>=1.0.4
google-genai